Added support for role adding
Bouncer can now provide a role when a user reacts to a message in a
specified channel. This is useful for forcing new members to verify
they are human by performing a simple task. New fields in config.json
are now required.

Also added front ends to some of the pre-existing debug functions.