Remove legacy DM handling code
Removes the old legacy DM handling code, in favor of the new
thread-based system. This commit should change no functionality from
when the 'createThreads' flag is set to true, as has been the case on
the production instance for a while now.

For reference, the removal of the legacy DM method does impact several
other systems:

- We no longer need to know the ID of the ban appeal server
- There is now only one AnsweringMachine instance
- '$reply ^' is no longer functional